calico: compress felix-bpf-gpl tarball using xz
In mimicking the behavior of upstream calico creating a tarball
of the felix bpf gpl components, a bug was duplicated that
named the tarball with a .tar.gz extension but didn't actually
compress the tarball. Upstream eventually fixed this as part
of their conversion to using xz compression on the tarball in
https://github.com/projectcalico/calico/pull/9364

Fix in the calico builds by converting to use xz. Busybox's version
of tar does not support the --use-compress-program=COMMAND option
from gnu tar that upstream used, so instead use the -J option to get
xz compression.

(This also was tripping up malcontent scans, but a fix for that landed
upstream in https://github.com/chainguard-dev/malcontent/pull/733 )

Signed-off-by: Steve Beattie <steve.beattie@chainguard.dev>